requirements.txt
sanic>=0.3.1
couchbase>=2.2.1
PyYAML>=3.12
